+++ hc/api/tests/test_bounce.py
@@ -49,3 +49,8 @@ class BounceTestCase(BaseTestCase):
         url = "/api/v1/notifications/%s/bounce" % fake_code
         r = self.client.post(url, "", content_type="text/plain")
         self.assertEqual(r.status_code, 404)
+
+    def test_it_requires_post(self):
+        url = "/api/v1/notifications/%s/bounce" % self.n.code
+        r = self.client.get(url)
+        self.assertEqual(r.status_code, 405)