+++ hc/front/views.py
@@ -712,7 +712,7 @@ def unsubscribe_email(request, code, token):
 
     # Some email servers open links in emails to check for malicious content.
     # To work around this, we serve a form that auto-submits with JS.
-    if "ask" in request.GET and request.method != "POST":
+    if request.method != "POST":
         return render(request, "accounts/unsubscribe_submit.html")
 
     channel.delete()